requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-pubsub==2.17.1
google-api-core==2.11.1
google-api-python-client==2.91.0

